perf: avoid frequently creating functions in cache.ensure
This mainly addresses the perf issue on large amount of calls to
`entry.new`. Previously every `cache.ensure` calls in the code path of
it creates an anonymous function, and it seems that luajit just could
not inline it. So finally this results in plenty of functions created
which results in high ram usage and needless overhead of checking cached
value. This commit solves this by ensuring exactly one function is
created for each cached call. From my testing, this has 3~5x perf
improvements on entry creation.